Code Index

Constructors

Descriptor

GTPSA.Descriptor — Type

Descriptor(nv::Integer, vo::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables of maximum order vo for each.

Input

nv – Number of variables in the TPSA

vo – Maximum order of the variables in the TPSA

Descriptor(vos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(mos) variables with individual truncation orders specified in the Vector vos.

Input

vos – Vector of the individual truncation orders of each variable

Descriptor(nv::Integer, vo::Integer, np::Integer, po::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables each with truncation order vo, and np parameters each with truncation order po

Input

nv – Number of variables in the TPSA

vo – Truncation order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Truncation order of the parameters

Descriptor(vos::Vector{<:Integer}, pos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(vos) variables with individual truncation orders specified in vos, and length(pos) parameters with individual truncation orders specified in pos.

Input

vos – Vector of the individual truncation orders of each variable

pos – Vector of the individual truncation orders of each parameter

TPS

GTPSA.TPS — Type

TPS(d::Descriptor)::TPS

Creates a new Truncated Power Series TPS based on d

Input

d – Descriptor

TPS(t1::TPS)::TPS

Copy constructor for TPS

Input

t1 – TPS to create new copy of

TPS(a::Real, t1::TPS)::TPS

Promotes the scalar a to a new TPS using the same Descriptor as t1

Input

a – Scalar to create new TPS with

t1 – TPS to use same Descriptor as

ComplexTPS

GTPSA.ComplexTPS — Type

ComplexTPS(d::Descriptor)::ComplexTPS

Creates a new Complex Truncated Power Series ComplexTPS based on d

Input

d – Descriptor

ComplexTPS(ct1::ComplexTPS)::ComplexTPS

Copy constructor for ComplexTPS

Input

ct1 – ComplexTPS to create new copy of

ComplexTPS(t1::TPS)::ComplexTPS

Creates a new copy of TPS promoted to a ComplexTPS

Input

t1 – TPS to create new ComplexTPS from

ComplexTPS(t1::TPS, t2::TPS)::ComplexTPS

Creates a new ComplexTPS equal to t1 + im*t2

Input

t1 – Real part of ComplexTPS as a TPS

t2 – Imaginary part of ComplexTPS as a TPS

ComplexTPS(a::Number, ct1::ComplexTPS)::ComplexTPS

Promotes the scalar a to a new ComplexTPS using the same Descriptor as ct1

Input

a – Scalar to create new ComplexTPS with

ct1 – ComplexTPS to use same Descriptor as

ComplexTPS(a::Number, t1::TPS)::ComplexTPS

Promotes the scalar a to a new ComplexTPS using the same Descriptor as t1

Input

a – Scalar to create new ComplexTPS with

t1 – TPS to use same Descriptor as

Monomial Getters/Setters

GTPSA.vars — Function

vars(d::Descriptor)::Vector{TPS}

Returns TPSs corresponding to the variables for the Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit TPSs corresponding to each variable

GTPSA.complexvars — Function

complexvars(d::Descriptor)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the variables for the Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit ComplexTPSs corresponding to each variable

GTPSA.params — Function

params(d::Descriptor)::Vector{TPS}

Returns TPSs corresponding to the parameters for the Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit TPSs corresponding to each parameter

GTPSA.complexparams — Function

complexparams(d::Descriptor)::Vector{TPS}

Returns ComplexTPSs corresponding to the parameters for the Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit ComplexTPSs corresponding to each parameter