Constructors

Descriptor

GTPSA.Descriptor — Type

Descriptor for TPSA, specifying the number of variables, number of parameters, orders of each variable, and orders of each parameters.

Fields

desc – Ptr to C Struct of the descriptor (Desc)

TPSA

GTPSA.TPSA — Type

TPSA containing only real coefficients

Fields

tpsa – Ptr to C Struct of the real TPSA (RTPSA)

ComplexTPSA

GTPSA.ComplexTPSA — Type

TPSA containing complex coefficients

Fields

tpsa – Ptr to C Struct of the complex TPSA (CTPSA)