ComplexTPS

Complex Truncated Power Series struct ComplexTPS <: Number

Syntax

ct = ComplexTPS([number] [, use=(descriptor|tps|complextps)])

ct = ComplexTPS([real1, real2] [, use=(descriptor|tps|complextps)])

Description

ct = ComplexTPS() creates a new ComplexTPS with all coefficients equal to zero using the Descriptor in GTPSA.desc_current

ct = ComplexTPS(number) creates a new ComplexTPS equal to number. If number is a TPS/ComplexTPS, then its Descriptor is used (equivalent to a copy constructor). Else, the Descriptor in GTPSA.desc_current is used

ct = ComplexTPS(real1, real2) creates a new ComplexTPS equal to complex(real1, real2). If real1 or real2 is a TPS then its Descriptor is used. Else, the Descriptor in GTPSA.desc_current is used

Optional Keyword Argument

use=descriptor creates a new ComplexTPS having a Descriptor equal to that passed

use=(tps|complextps) creates a new ComplexTPS having a Descriptor equal to that used by the passed TPS or ComplexTPS

Examples

julia> GTPSA.show_header = truetrue
julia> d1 = Descriptor(1, 1); # 1 variable to order 1
julia> t1_1 = ComplexTPS()ComplexTPS:
-----------------------
# Variables:       1
Variable order:    1
-----------------------
 Real                     Imag                       Order   Exponent
  0.0000000000000000e+00   0.0000000000000000e+00      0      0
julia> t2_1 = ComplexTPS(5)ComplexTPS:
-----------------------
# Variables:       1
Variable order:    1
-----------------------
 Real                     Imag                       Order   Exponent
  5.0000000000000000e+00   0.0000000000000000e+00      0      0
julia> t3_1 = ComplexTPS(6, TPS(5))ComplexTPS:
-----------------------
# Variables:       1
Variable order:    1
-----------------------
 Real                     Imag                       Order   Exponent
  6.0000000000000000e+00   5.0000000000000000e+00      0      0
julia> t3_1 = ComplexTPS(TPS(5))ComplexTPS:
-----------------------
# Variables:       1
Variable order:    1
-----------------------
 Real                     Imag                       Order   Exponent
  5.0000000000000000e+00   0.0000000000000000e+00      0      0
julia> d2 = Descriptor(1, 10); # New Descriptor to order 10
julia> t3_2 = ComplexTPS(t3_1, use=d2) # Copies and changes DescriptorComplexTPS:
-----------------------
# Variables:       1
Variable order:    10
-----------------------
 Real                     Imag                       Order   Exponent
  5.0000000000000000e+00   0.0000000000000000e+00      0      0

Documentation

GTPSA.ComplexTPS — Type

ComplexTPS(cta::Union{Number,Nothing}=nothing; use::Union{Descriptor,TPS,ComplexTPS,Nothing}=nothing)::ComplexTPS

Constructor to create a new ComplexTPS equal to the number cta. If cta is a ComplexTPS (or TPS), this is equivalent to a copy constructor, with the result by default having the same Descriptor as cta. If cta is not a TPS orComplexTPS, then the Descriptor used will by default be GTPSA.desc_current. The Descriptor for the constructed ComplexTPS can be set using use. If a TPS or ComplexTPS is passed to use, the Descriptor for that TPS will be used.

The constructor can also be used to create a copy of a ComplexTPS under one Descriptor to instead have a different Descriptor. In this case, invalid monomials under the new Descriptor are removed.

Input

cta – Any Number

use – (Optional) specify which Descriptor to use, default is nothing which uses the Descriptor for cta if cta <: Union{TPS,ComplexTPS}, else uses GTPSA.desc_current

Output

ret – New ComplexTPS equal to cta with removal of invalid monomials if cta is a TPS/ComplexTPS and a new Descriptor is specified