vars, params

Creates a vector of TPSs corresponding to each variable/parameter in the GTPSA

Syntax

# Dynamic Descriptor Resolution:
Δx  = vars([descriptor|tps])
Δk  = params([descriptor|tps])
Δx  = complexvars([descriptor|tps])
Δk  = complexparams([descriptor|tps])

# Static Descriptor Resolution:
Δx  = @vars(descriptor [,complex=bool])
Δk  = @params(descriptor [,complex=bool])

Dynamic Descriptor Resolution

Δx = vars([descriptor|tps]) creates a vector of TPSs corresponding to each of the variables in the GTPSA given by descriptor|tps, which defaults to GTPSA.desc_current if not provided.

Δk = params([descriptor|tps]) creates a vector of TPSs corresponding to each of the parameters in the GTPSA given by descriptor|tps, which defaults to GTPSA.desc_current if not provided

Δx = complexvars([descriptor|tps]) creates a vector of complex TPSs corresponding to each of the variables in the GTPSA given by descriptor|tps, which defaults to GTPSA.desc_current if not provided.

Δk = complexparams([descriptor|tps]) creates a vector of complex TPSs corresponding to each of the parameters in the GTPSA given by descriptor|tps, which defaults to GTPSA.desc_current if not provided

Examples

julia> d5 = Descriptor(3, 5, 2, 5); # 3 vars, 2 params, all to order 5
julia> Δx = vars(d5)3-element Vector{TPS64{GTPSA.Dynamic}}:
Descriptor(NV=3, MO=5, NP=2, PO=5)
INDEX  COEFFICIENT             ORDER   EXPONENTS
------------------------------------------------
  1:   1.0000000000000000E+00    1     1 0  0
------------------------------------------------
  2:   1.0000000000000000E+00    1     0 1  0
------------------------------------------------
  3:   1.0000000000000000E+00    1     0 0  1
julia> Δxc = complexvars(d5)3-element Vector{ComplexTPS64{GTPSA.Dynamic}}:
Descriptor(NV=3, MO=5, NP=2, PO=5)
INDEX  COEFFICIENT             ORDER   EXPONENTS
----------------------------------------------------------------------
  1:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     1 0  0
----------------------------------------------------------------------
  2:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 1  0
----------------------------------------------------------------------
  3:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  1
julia> Δkc = complexparams(d5)2-element Vector{ComplexTPS64{GTPSA.Dynamic}}:
Descriptor(NV=3, MO=5, NP=2, PO=5)
INDEX  COEFFICIENT             ORDER   EXPONENTS
---------------------------------------------------------------------------
  1:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  0  4^1
---------------------------------------------------------------------------
  2:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  0  5^1

Static Descriptor Resolution

Δx = @vars(descriptor) creates a vector of TPSs corresponding to each of the variables in the GTPSA descriptor

Δk = @params(descriptor) creates a vector of TPSs corresponding to each of the parameters in the GTPSA descriptor

Optional Argument

complex if true, will return the corresponding TPSs as ComplexTPS64. Default is false.

Examples

julia> d5 = Descriptor(3, 5, 2, 5); # 3 vars, 2 params, all to order 5
julia> Δx = @vars(d5)3-element Vector{TPS64{Descriptor(NV=3, MO=5, NP=2, PO=5)}}:
INDEX  COEFFICIENT             ORDER   EXPONENTS
------------------------------------------------
  1:   1.0000000000000000E+00    1     1 0  0
------------------------------------------------
  2:   1.0000000000000000E+00    1     0 1  0
------------------------------------------------
  3:   1.0000000000000000E+00    1     0 0  1
julia> Δxc = @vars(d5, complex=true)3-element Vector{ComplexTPS64{Descriptor(NV=3, MO=5, NP=2, PO=5)}}:
INDEX  COEFFICIENT             ORDER   EXPONENTS
----------------------------------------------------------------------
  1:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     1 0  0
----------------------------------------------------------------------
  2:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 1  0
----------------------------------------------------------------------
  3:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  1
julia> Δk = @params(d5, complex=true)2-element Vector{ComplexTPS64{Descriptor(NV=3, MO=5, NP=2, PO=5)}}:
INDEX  COEFFICIENT             ORDER   EXPONENTS
---------------------------------------------------------------------------
  1:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  0  4^1
---------------------------------------------------------------------------
  2:   1.0000000000000000E+00 +0.0000000000000000E+00i    1     0 0  0  5^1

Documentation

GTPSA.vars — Function

vars(use::Union{Descriptor,TPS}=GTPSA.desc_current)

Returns a vector of TPSs corresponding to the variables for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit TPS{Float64}s corresponding to each variable

GTPSA.params — Function

params(use::Union{Descriptor,TPS}=GTPSA.desc_current)

Returns a vector of TPSs corresponding to the parameters for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit TPS{Float64}s corresponding to each parameters

GTPSA.complexvars — Function

complexvars(use::Union{Descriptor,TPS}=GTPSA.desc_current)

Returns a vector of ComplexTPS64s corresponding to the variables for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit ComplexTPS64s corresponding to each variable

GTPSA.complexparams — Function

complexparams(use::Union{Descriptor,TPS}=GTPSA.desc_current)

Returns a vector of ComplexTPS64s corresponding to the parameters for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit ComplexTPS64s corresponding to each parameters

GTPSA.@vars — Macro

@vars(descriptor [,complex=bool])

Constructs a vector of TPSs corresponding to each of the variables in the GTPSA descriptor

Keyword Arguments

complex – If true, returns the variables as ComplexTPS64s. Default is false.

GTPSA.@params — Macro

@params(descriptor [,complex=bool])

Constructs a vector of TPSs corresponding to each of the parameters in the GTPSA descriptor

Keyword Arguments

complex – If true, returns the parameters as ComplexTPS64s. Default is false.