Descriptor

Defines the number of variables, number of parameters, and order(s) for each in the GTPSA

Syntax

d = Descriptor(num_vars, var_order)     
d = Descriptor(var_orders)
d = Descriptor(num_vars, var_order, num_params, param_order)   
d = Descriptor(var_orders, param_orders)  

Description

d = Descriptor(num_vars, var_order) defines a GTPSA Descriptor with num_vars variables all having the same truncation order var_order

d = Descriptor(var_orders) defines a GTPSA Descriptor with length(var_orders) variables each having individual truncation orders specified in the var_orders vector

d = Descriptor(num_vars, var_order, num_params, param_order) defines a GTPSA Descriptor with num_vars variables all having the same truncation order var_order, and num_params parameters all having the same truncation order param_order

d = Descriptor(var_orders, param_orders) defines a GTPSA Descriptor with length(var_orders) variables each having individual truncation orders specified in the var_orders vector, and length(param_orders) parameters each having individual truncation orders specified in the param_orders vector

A Descriptor defines all information about the GTPSA, including the number of variables and parameters, and the truncation orders for each variable and parameter. Before any operations using TPSs, a Descriptor must be defined. Each time a new Descriptor is created, the non-constant global variable GTPSA.desc_current is set. If a Descriptor is not explicitly (or implicitly in a TPS copy constructor) passed to the TPS Constructors, GTPSA.desc_current will be used by default. This non-constant global can be set by the user throughout a program using multiple Descriptors.

Examples

julia> d1 = Descriptor(2, 10)GTPSA Descriptor
-----------------------
# Variables:       2
Variable order:    10
julia> d2 = Descriptor([1, 2, 3])GTPSA Descriptor
-----------------------
# Variables:       3
Variable orders:  [1, 2, 3]
julia> d3 = Descriptor(3, 4, 1, 2)GTPSA Descriptor
-----------------------
# Variables:       3
Variable order:    4
# Parameters:      1
Parameter order:   2
julia> d4 = Descriptor([6, 5], [4])GTPSA Descriptor
-----------------------
# Variables:       2
Variable orders:  [6, 5]
# Parameters:      1
Parameter order:   4
julia> GTPSA.desc_current = d1GTPSA Descriptor
-----------------------
# Variables:       2
Variable order:    10

Documentation

GTPSA.Descriptor — Type

Descriptor(nv::Integer, vo::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables of maximum order vo for each.

Input

nv – Number of variables in the TPSA

vo – Maximum order of the variables in the TPSA

Descriptor(vos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(mos) variables with individual truncation orders specified in the Vector vos.

Input

vos – Vector of the individual truncation orders of each variable

Descriptor(nv::Integer, vo::Integer, np::Integer, po::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables each with truncation order vo, and np parameters each with truncation order po

Input

nv – Number of variables in the TPSA

vo – Truncation order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Truncation order of the parameters

Descriptor(vos::Vector{<:Integer}, pos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(vos) variables with individual truncation orders specified in vos, and length(pos) parameters with individual truncation orders specified in pos.

Input

vos – Vector of the individual truncation orders of each variable

pos – Vector of the individual truncation orders of each parameter