vars, params

Creates a vector of TPSs corresponding to each variable/parameter in the GTPSA

Syntax

x  = vars([(descriptor|tps|complextps)])
xc = complexvars([(descriptor|tps|complextps)])

k  = params([(descriptor|tps|complextps)])
kc = complexparams([(descriptor|tps|complextps)])

Description

x = vars() creates a vector of TPSs corresponding to each of the variables in the GTPSA defined by the Descriptor in GTPSA.desc_current

x = complexvars() creates a vector of ComplexTPSs corresponding to each of the variables in the GTPSA defined by the Descriptor in GTPSA.desc_current

k = params() creates a vector of TPSs corresponding to each of the parameters in the GTPSA defined by the Descriptor in GTPSA.desc_current

k = complexparams() creates a vector of ComplexTPSs corresponding to each of the parameters in the GTPSA defined by the Descriptor in GTPSA.desc_current

Optional Argument

descriptor creates a vector of TPS/ComplexTPSs corresponding to each of the variables/parameters in the GTPSA defined by the passed Descriptor

(tps|complextps) creates a vector of TPS/ComplexTPSs corresponding to each of the variables/parameters in the GTPSA defined by the Descriptor of the passed TPS or ComplexTPS

Examples

julia> GTPSA.show_header=truetrue
julia> d1 = Descriptor(3, 5, 2, 5); # 3 vars, 2 params, all to order 5
julia> x1 = vars()3-element Vector{TPS}:
-----------------------
Maximum order:     5
# Variables:       3
Parameter order:   5
# Parameters:      2
-----------------------
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   0   |   0   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   0   |   0   0
-----------------------------------------------------------------
   3:   1.0000000000000000e+00      1      0   0   1   |   0   0
julia> k1 = params()2-element Vector{TPS}:
-----------------------
Maximum order:     5
# Variables:       3
Parameter order:   5
# Parameters:      2
-----------------------
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   0   |   1   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   0   0   |   0   1
julia> d2 = Descriptor(2, 5, 1, 5); # 2 vars, 1 param, all to order 5
julia> x2 = vars()2-element Vector{TPS}:
-----------------------
Maximum order:     5
# Variables:       2
Parameter order:   5
# Parameters:      1
-----------------------
  Out  Coefficient                Order   Exponent
---------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   |   0
---------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   |   0
julia> k2 = params()1-element Vector{TPS}:
-----------------------
Maximum order:     5
# Variables:       2
Parameter order:   5
# Parameters:      1
-----------------------
  Out  Coefficient                Order   Exponent
---------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   |   1
julia> k1 = params(d1)2-element Vector{TPS}:
-----------------------
Maximum order:     5
# Variables:       3
Parameter order:   5
# Parameters:      2
-----------------------
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   0   |   1   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   0   0   |   0   1

Documentation

GTPSA.vars — Function

vars(use::Union{Descriptor,TPS,ComplexTPS}=GTPSA.desc_current)::Vector{TPS}

Returns TPSs corresponding to the variables for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which TPSA Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit TPSs corresponding to each variable

GTPSA.complexvars — Function

complexvars(use::Union{Descriptor,TPS,ComplexTPS}=GTPSA.desc_current)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the variables for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which TPSA Descriptor to use, default is GTPSA.desc_current

Output

x – Vector containing unit ComplexTPSs corresponding to each variable

GTPSA.params — Function

params(use::Union{Descriptor,TPS,ComplexTPS}=GTPSA.desc_current)::Vector{TPS}

Returns TPSs corresponding to the parameters for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which TPSA Descriptor to use, default is GTPSA.desc_current

Output

k – Vector containing unit TPSs corresponding to each parameter

GTPSA.complexparams — Function

complexparams(d::Descriptor=GTPSA.desc_current)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the parameters for the Descriptor specified by use. Default value is GTPSA.desc_current.

Input

use – (Optional) Specify which TPSA Descriptor to use, default is GTPSA.desc_current

Output

k – Vector containing unit ComplexTPSs corresponding to each parameter