gradient, jacobian, hessian

Extracts specific partial derivatives from a TPS

Syntax

grad = gradient(f [, include_params=bool])
gradient!(grad, f [, include_params=bool])

J = jacobian(F [, include_params=bool])
jacobian!(J, F [, include_params=bool])

Jt = jacobiant(F [, include_params=bool])
jacobiant!(Jt, F [, include_params=bool])

H = hessian(f [, include_params=bool])
hessian!(H, f [, include_params=bool])

Description

grad = gradient(f) extracts the gradient from the TPS f, defined as $\nabla f$

gradient!(grad, f) fills grad vector in-place with the gradient extracted from the TPS f

J = jacobian(F) extracts the Jacobian matrix from the vector of TPSs F, defined as $J_{ij} = \frac{\partial F_i}{\partial x_j}$

jacobian!(J, F) fills the J matrix in-place with the Jacobian extracted from the vector of TPSs F

Jt = jacobiant(F) extracts the transpose of the Jacobian matrix from the vector of TPSs F, with the Jacobian defined as $J_{ij} = \frac{\partial F_i}{\partial x_j}$

jacobiant!(Jt, F) fills the Jt matrix in-place with the transpose of the Jacobian extracted from the vector of TPSs F

H = hessian(f) extracts the Hessian matrix from the TPS f, defined as $H_{ij} = \frac{\partial^2 f}{\partial x_i\partial x_j}$

hessian!(H, f) fills the H matrix in-place with the Hessian extracted from the TPS f

Optional Argument

include_params can be set to true (default is false) to include the partial derivatives with respect to the parameters in any of the extraction functions above.

Examples

julia> d = Descriptor(2,10);
julia> x = vars(d);
julia> f = x[1] + 2*x[2] + 3*x[1]^2 + 4*x[1]*x[2] + 5*x[2]^2;
julia> g = 5*x[1] + 4*x[2] + 3*x[1]^2 + 2*x[1]*x[2] + x[2]^2;
julia> grad = gradient(f)2-element Vector{Float64}:
 1.0
 2.0
julia> J = jacobian([f, g])2×2 Matrix{Float64}:
 1.0  2.0
 5.0  4.0
julia> H = hessian(f)2×2 Matrix{Float64}:
 6.0   4.0
 4.0  10.0

Documentation

GTPSA.gradient — Function

gradient(t::Union{TPS,ComplexTPS}; include_params=false)

Extracts the first-order partial derivatives (evaluated at 0) from the TPS. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPS.

Input

t – TPS/ComplexTPS to extract the gradient from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

grad – Gradient of the TPS

GTPSA.gradient! — Function

gradient!(result::Vector{<:Union{Float64,ComplexF64}}, t::Union{TPS,ComplexTPS}; include_params=false)

Extracts the first-order partial derivatives (evaluated at 0) from the TPS and fills the result vector in-place. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPS.

Input

t – TPS/ComplexTPS to extract the gradient from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

result – Preallocated Vector to fill with the gradient of the TPS

GTPSA.jacobian — Function

jacobian(m::Union{Vector{<:Union{TPS,ComplexTPS}}, SubArray{<:Union{TPS,ComplexTPS},N,P,I,true}}; include_params=false) where {N,P,I}

Extracts the first-order partial derivatives (evaluated at 0) from the Vector of TPSs. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPSs.

Input

m – Vector of TPSs. to extract the Jacobian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

J – Jacobian of m

GTPSA.jacobian! — Function

jacobian!(result::Matrix{<:Union{Float64,ComplexF64}}, m::Union{Vector{<:Union{TPS,ComplexTPS}}, SubArray{<:Union{TPS,ComplexTPS},N,P,I,true}}; include_params=false)  where {N,P,I}

Extracts the first-order partial derivatives (evaluated at 0) from the Vector of TPSs. and fills the result matrix in-place. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPSs.

Input

m – Vector of TPSs. to extract the Jacobian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

result – Preallocated matrix to fill with the Jacobian of m

GTPSA.jacobiant — Function

jacobiant(m::Union{Vector{<:Union{TPS,ComplexTPS}}, SubArray{<:Union{TPS,ComplexTPS},N,P,I,true}}; include_params=false) where {N,P,I}

Extracts the first-order partial derivatives (evaluated at 0) from the Vector of TPSs, as the transpose of the Jacobian. Because of Julia's column-major indexing vs. C's row-major, this routine will be slightly faster than using jacobian if the transpose of the Jacobian is needed. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPSs.

Input

m – Vector of TPSs. to extract the Jacobian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

Jt – Transpose of the Jacobian of m

GTPSA.jacobiant! — Function

jacobiant!(result::Matrix{<:Union{Float64,ComplexF64}}, m::Union{Vector{<:Union{TPS,ComplexTPS}}, SubArray{<:Union{TPS,ComplexTPS},N,P,I,true}}; include_params=false) where {N,P,I}

Extracts the first-order partial derivatives (evaluated at 0) from the Vector of TPSs, as the transpose of the Jacobian. Because of Julia's column-major indexing vs. C's row-major, this routine will be slightly faster than using jacobian! if the transpose of the Jacobian is needed. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the first-order monomial coefficients already in the TPSs and filling result.

Input

m – Vector of TPSs. to extract the Jacobian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

result – Preallocated matrix to fill with the transpose of the Jacobian of m

GTPSA.hessian — Function

hessian(t::Union{TPS,ComplexTPS}; include_params=false)

Extracts the second-order partial derivatives (evaluated at 0) from the TPS. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the second-order monomial coefficients already in the TPS.

Input

t – TPS/ComplexTPS to extract the Hessian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

H – Hessian of the TPS

GTPSA.hessian! — Function

hessian!(result::Matrix{<:Union{Float64,ComplexF64}},t::Union{TPS,ComplexTPS}; include_params=false)

Extracts the second-order partial derivatives (evaluated at 0) from the TPS and fills the result matrix in-place. The partial derivatives wrt the parameters will also be extracted when the include_params flag is set to true. Note that this function is not calculating anything - just extracting the second-order monomial coefficients already in the TPS.

Input

t – TPS/ComplexTPS to extract the Hessian from

include_params – (Optional) Extract partial derivatives wrt parameters. Default is false

Output

result – Preallocated matrix to fill with the Hessian of the TPS