vars, params

Creates a vector of TPSs corresponding to each variable/parameter in the GTPSA

Syntax

x = vars(descriptor)
x = vars()
x = complexvars(descriptor)
x = complexvars()

k = params(descriptor)
k = params()
k = complexparams(descriptor)
k = complexparams()

Description

x = vars(descriptor) creates a vector of TPSs corresponding to each of the variables in the GTPSA defined by the passed Descriptor

x = vars() creates a vector of TPSs corresponding to each of the variables in the GTPSA defined by the Descriptor in GTPSA.desc_current

x = complexvars(descriptor) creates a vector of ComplexTPSs corresponding to each of the variables in the GTPSA defined by the passed Descriptor

x = complexvars() creates a vector of ComplexTPSs corresponding to each of the variables in the GTPSA defined by the Descriptor in GTPSA.desc_current

k = params(descriptor) creates a vector of TPSs corresponding to each of the parameters in the GTPSA defined by the passed Descriptor

k = params() creates a vector of TPSs corresponding to each of the parameters in the GTPSA defined by the Descriptor in GTPSA.desc_current

k = complexparams(descriptor) creates a vector of ComplexTPSs corresponding to each of the parameters in the GTPSA defined by the passed Descriptor

k = complexparams() creates a vector of ComplexTPSs corresponding to each of the parameters in the GTPSA defined by the Descriptor in GTPSA.desc_current

Examples

julia> d1 = Descriptor(3, 5, 2, 5); # 3 vars, 2 params, all to order 5
julia> x1 = vars()3-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   0   |   0   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   0   |   0   0
-----------------------------------------------------------------
   3:   1.0000000000000000e+00      1      0   0   1   |   0   0
julia> k1 = params()2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   0   |   1   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   0   0   |   0   1
julia> d2 = Descriptor(2, 5, 1, 5); # 2 vars, 1 param, all to order 5
julia> x2 = vars()2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
---------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   |   0
---------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   |   0
julia> k2 = params()1-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
---------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   |   1
julia> k1 = params(d1)2-element Vector{TPS}:
  Out  Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      0   0   0   |   1   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   0   0   |   0   1

Documentation

GTPSA.vars — Function

vars(d::Descriptor=GTPSA.desc_current)::Vector{TPS}

Returns TPSs corresponding to the variables for the Descriptor. Default value is the most recently-defined Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit TPSs corresponding to each variable

GTPSA.complexvars — Function

complexvars(d::Descriptor=GTPSA.desc_current)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the variables for the Descriptor. Default value is the most recently-defined Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit ComplexTPSs corresponding to each variable

GTPSA.params — Function

params(d::Descriptor=GTPSA.desc_current)::Vector{TPS}

Returns TPSs corresponding to the parameters for the Descriptor. Default value is the most recently-defined Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit TPSs corresponding to each parameter

GTPSA.complexparams — Function

complexparams(d::Descriptor=GTPSA.desc_current)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the parameters for the Descriptor. Default value is the most recently-defined Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit ComplexTPSs corresponding to each parameter