GTPSA Structs

Descriptor

GTPSA.Descriptor — Type

Descriptor(nv::Integer, mo::Integer)

Creates a TPSA Descriptor with nv variables of maximum order mo.

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

Descriptor(nv::Integer, mo::Integer, np::Integer, po::Integer)

Creates a TPSA Descriptor with nv variables of maximum order mo, and np parameters of maximum order po (<= mo).

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Maximum order of the parameters (<= mo) in the TPSA

Descriptor(nv::Integer, mo::Integer, np::Integer, po::Integer, no::Vector{Int})

Creates a TPSA Descriptor with nv variables of maximum order mo, np parameters of maximum order po (<= mo), and the individual variable/parameter orders specified in the Vector no.

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Maximum order of the parameters (<= mo) in the TPSA

no – Vector of the individual orders of each variable and parameter, with variables first nv entries and parameters last np entries

TPSA

GTPSA.TPSA — Type

TPSA()

Creates a TPSA using the most recently-defined Descriptor

TPSA(d::Descriptor)

Creates a TPSA based on d

Input

d – Descriptor

TPSA(t1::TPSA)

Creates a TPSA with the same Descriptor as t1

Input

t1 – TPSA to create new TPSA from with same Descriptor

TPSA(t1::AbstractTPSA)

Creates a TPSA with the same Descriptor as t1

Input

t1 – TPSA to create new TPSA from with same Descriptor

ComplexTPSA

GTPSA.ComplexTPSA — Type

ComplexTPSA()

Creates a ComplexTPSA using the most recently-defined Descriptor

ComplexTPSA(d::Descriptor)

Creates a ComplexTPSA based on d

Input

d – Descriptor

ComplexTPSA(t1::ComplexTPSA)

Creates a ComplexTPSA with the same Descriptor as t1

Input

t1 – TPSA to create new TPSA from with same Descriptor

ComplexTPSA(t1::AbstractTPSA)

Creates a ComplexTPSA with the same Descriptor as t1

Input

t1 – TPSA to create new TPSA from with same Descriptor