Usage

In some use cases, the TPS may define the evolution of some state $x$ of a dynamical system, e.g. $x_f = \mathcal{M}(x_i,k_1)$, where $k_1$ is some defined external parameter. When $\mathcal{M}$ is expanded in powers of $x_i$, this is referred to as a Taylor map. $x$ is referred to as a map variable or variable, as its evolution is defined by the TPS. $k_1$ which does not evolve with $\mathcal{M}$, is referred to as a parameter. GTPSA allows distinction between map variables and parameters, as well as the individual orders for each, to speed up computations.

Custom Variable/Parameter Orders

The Descriptor defines all information about the TPSA, including the number of variables, order for each variable, number of parameters, and order for each parameter. The constructors for a Descriptor are:

# Descriptor for 2 variables with max order 10 for each
d1 = Descriptor(2, 10)    # Max order 10 for all variables
d1 = Descriptor([10, 10]) # Equivalent definition specifying order for each

# Descriptor for 3 variables with max order 4 for each, and 1 parameter with max order 2
d2 = Descriptor(3, 4, 1, 2)     # Max order 4 for all variables, 2 for all parameters
d2 = Descriptor([4, 4, 4], [2]) # Equivalent definition specifying order for each

# Descriptor for 2 variables with max orders 6, 5 respectively and 1 parameter with max order 4
d3 = Descriptor([6, 5], [4])

The constructor for Descriptor definitions are:

GTPSA.Descriptor — Type

Descriptor(nv::Integer, mo::Integer)

Creates a TPSA Descriptor with nv variables of maximum order mo.

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

Descriptor(mos::Vector{<:Integer})

Creates a TPSA Descriptor with length(mos) variables with individual max orders specified in the Vector mos.

Input

mos – Vector of the individual max orders of each variable

Descriptor(nv::Integer, mo::Integer, np::Integer, po::Integer)

Creates a TPSA Descriptor with nv variables of maximum order mo, and np parameters of maximum order po (<= mo).

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Maximum order of the parameters (<= mo) in the TPSA

Descriptor(mos::Vector{<:Integer}, pos::Vector{<:Integer})

Creates a TPSA Descriptor with length(mos) variables with individual max orders specified in mos, and length(pos) parameters with individual max orders specified in pos.

Input

mos – Vector of the individual max orders of each variable

pos – Vector of the individual max orders of each parameter

TPS Constructors

After defining a Descriptor for the TPSA, generally one would want the variables or parameters in the descriptor, defined as TPSs or ComplexTPSs. The convenience functions vars and params, and their complex counterparts complexvars and complexparams are provided which return a Vector of all variables in the TPSA as TPSs (or ComplexTPSs). For example, suppose we wish to evaluate the Taylor series for $f(x_1,x_2,k_1) = \cos{(x_1)} + \sin{(k_1)}\sqrt{1+x_2}$ to 10th order in $x_1$ and $x_2$ but only 5th order in the parameter $k_1$:

d = Descriptor(2, 10)

x = vars(d)     # Returns a Vector of each variable as a TPS
k = params(d)   # Returns a Vector of each parameter as a TPS

# f is a TPS containing the result
f = cos(x[1]) + sin(k[1]) * sqrt(1 + x[2])

A blank TPS or ComplexTPS with all coefficients equal to zero can be created using TPS(d) or ComplexTPS(d) respectively.

TPS

GTPSA.vars — Function

vars(d::Descriptor)::Vector{TPS}

Returns TPSs corresponding to the variables for the Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit TPSs corresponding to each variable

GTPSA.params — Function

params(d::Descriptor)::Vector{TPS}

Returns TPSs corresponding to the parameters for the Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit TPSs corresponding to each parameter

GTPSA.TPS — Type

TPS()

Creates a Truncated Power Series TPS using the most recently-defined Descriptor

TPS(d::Descriptor)

Creates a Truncated Power Series TPS based on d

Input

d – Descriptor

TPS(t1::TPS)

Creates a Truncated Power Series TPS with the same Descriptor as t1

Input

t1 – TPS to create new TPS from with same Descriptor

TPS(t1::AbstractTPS)

Creates a Truncated Power Series TPS with the same Descriptor as t1

Input

t1 – TPS to create new TPS from with same Descriptor

ComplexTPS

GTPSA.complexvars — Function

complexvars(d::Descriptor)::Vector{ComplexTPS}

Returns ComplexTPSs corresponding to the variables for the Descriptor.

Input

d – TPSA Descriptor

Output

x – Vector containing unit ComplexTPSs corresponding to each variable

GTPSA.complexparams — Function

complexparams(d::Descriptor)::Vector{TPS}

Returns ComplexTPSs corresponding to the parameters for the Descriptor.

Input

d – TPSA Descriptor

Output

k – Vector containing unit ComplexTPSs corresponding to each parameter

GTPSA.ComplexTPS — Type

ComplexTPS()

Creates a Complex Truncated Power Series ComplexTPS using the most recently-defined Descriptor

ComplexTPS(d::Descriptor)

Creates a Complex Truncated Power Series ComplexTPS based on d

Input

d – Descriptor

ComplexTPS(t1::ComplexTPS)

Creates a Complex Truncated Power Series ComplexTPS with the same Descriptor as t1

Input

t1 – TPS to create new TPS from with same Descriptor

ComplexTPS(t1::AbstractTPS)

Creates a Complex Truncated Power Series ComplexTPS with the same Descriptor as t1

Input

t1 – TPS to create new TPS from with same Descriptor

Monomial Indexing

A TPS contains the Taylor coefficients for all of the monomials in the Taylor series up to the specified truncation order. The monomial coefficients in a TPS can be get/set with two methods of indexing:

By Order: A particular monomial in a TPS t can be indexed by t[<x_1 order>, ..., <x_nv order>, <k_1 order>, ..., <k_np order>]. For example, for a TPS with variables $x_1$, $x_2$ and parameter $k_1$, the $x_1^3x_2^1k_1^2$ monomial is accessed with t[3,1,2]. The 0th order part (the scalar part) of the TPS is set with t[0,0,0] or equivalently t[0] as leaving out indices for variables/parameters is allowed if they are not included.

# Example of indexing by order

# Descriptor for 2 variables with max order 15
d = Descriptor(2, 15)
x = vars(d)
x1 = x[1]
x2 = x[2]

y1 = TPS(d)         # Create blank TPS with zero for all coefficients
y1[1,0] = 1         # Set first-order part for first variable equal to 1

y1[1,0] == x1[1,0]  # Is true
y1 == x1            # Is true

y2 = TPS(d)         # Create blank TPS with zero for all coefficients
y2[0,1] = 1         # Set first-order part for second variable equal to 1

y2[0,1] == x2[0,1]  # Is true
y2 == x2            # Is true

By Var => Order, Param => Order: A particular monomial in a TPS t can be indexed by t[<ix_var> => <order>, ..., params=(<ix_param> => <order>, ...)]. This method of indexing is convenient when a TPS contains many variables and parameters. For example, for a TPS with variables $x_1$, $x_2$ and parameter $k_1$, the $x_1^3x_2^1k_1^2$ monomial is accessed with t[1=>3, 2=>1, params=(1=>2,)]. The 0th order part (the scalar part) of the TPS cannot be set with this method.

# Example of indexing by var => order, param => order

# Descriptor for 3 variables with order 5, and 2 parameters with orders 3, 4 respectively
d = Descriptor([5, 5, 5], [3, 4])
x = vars(d)
x1 = x[1]
x2 = x[2]
x3 = x[3]
k = params(d)
k1 = k[1]
k2 = k[2]

y1 = TPS(d)             # Create blank TPS with zero for all coefficients
y1[1=>1] = 1            # Set first-order part for first variable equal to 1     

y1 == x1                # Is true          

z1 = TPS(d)             # Create blank TPS with zero for all coefficients
z1[params=(1=>1,)] = 1  # Set first-order part for first parameter equal to 1

z1 == k1                # Is true

Using Complex TPSs

Mixing ComplexTPSs with TPSs is currently not allowed, however this may be added in the future. The usage for ComplexTPSs is the same as with regular TPSs, with the addition of some extra functions including real and imag.