Constructors

Descriptor

GTPSA.Descriptor — Type

Descriptor(nv::Integer, mo::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables of maximum order mo.

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

Descriptor(mos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(mos) variables with individual max orders specified in the Vector mos.

Input

mos – Vector of the individual max orders of each variable

Descriptor(nv::Integer, mo::Integer, np::Integer, po::Integer)::Descriptor

Creates a TPSA Descriptor with nv variables of maximum order mo, and np parameters of maximum order po (<= mo).

Input

nv – Number of variables in the TPSA

mo – Maximum order of the variables in the TPSA

np – Number of parameters in the TPSA

po – Maximum order of the parameters (<= mo) in the TPSA

Descriptor(mos::Vector{<:Integer}, pos::Vector{<:Integer})::Descriptor

Creates a TPSA Descriptor with length(mos) variables with individual max orders specified in mos, and length(pos) parameters with individual max orders specified in pos.

Input

mos – Vector of the individual max orders of each variable

pos – Vector of the individual max orders of each parameter

TPS

GTPSA.TPS — Type

TPS(d::Descriptor)::TPS

Creates a new Truncated Power Series TPS based on d

Input

d – Descriptor

TPS(t1::TPS)::TPS

Copy constructor for TPS

Input

t1 – TPS to create new copy of

TPS(a::Real, t1::TPS)::TPS

Promotes the scalar a to a new TPS using the same Descriptor as t1

Input

a – Scalar to create new TPS with

t1 – TPS to use same Descriptor as

ComplexTPS

GTPSA.ComplexTPS — Type

ComplexTPS(d::Descriptor)::ComplexTPS

Creates a new Complex Truncated Power Series ComplexTPS based on d

Input

d – Descriptor

ComplexTPS(ct1::ComplexTPS)::ComplexTPS

Copy constructor for ComplexTPS

Input

ct1 – ComplexTPS to create new copy of

ComplexTPS(t1::TPS)::ComplexTPS

Creates a new copy of TPS promoted to a ComplexTPS

Input

t1 – TPS to create new ComplexTPS from

ComplexTPS(t1::TPS, t2::TPS)::ComplexTPS

Creates a new ComplexTPS equal to t1 + im*t2

Input

t1 – Real part of ComplexTPS as a TPS

t2 – Imaginary part of ComplexTPS as a TPS

ComplexTPS(a::Number, ct1::ComplexTPS)::ComplexTPS

Promotes the scalar a to a new ComplexTPS using the same Descriptor as ct1

Input

a – Scalar to create new ComplexTPS with

ct1 – ComplexTPS to use same Descriptor as

ComplexTPS(a::Number, t1::TPS)::ComplexTPS

Promotes the scalar a to a new ComplexTPS using the same Descriptor as t1

Input

a – Scalar to create new ComplexTPS with

t1 – TPS to use same Descriptor as