@vars, @params

Creates a vector of TPSs corresponding to each variable/parameter in the GTPSA

Syntax

Δx  = @vars(descriptor [,complex=bool] [, dynamic=bool])

Δk  = @params(descriptor [,complex=bool] [, dynamic=bool])

Description

Δx = @vars(descriptor) creates a vector of TPSs corresponding to each of the variables in the GTPSA descriptor

Δk = params(descriptor) creates a vector of TPSs corresponding to each of the parameters in the GTPSA descriptor`

Optional Argument

complex if true, will return the corresponding TPSs as ComplexTPS64. Default is false.

dynamic if true, will use dynamic Descriptor resolution for the returned TPSs

Examples

julia> d5 = Descriptor(3, 5, 2, 5); # 3 vars, 2 params, all to order 5
julia> Δx = @vars(d5)3-element Vector{TPS64{Descriptor(NV=3, MO=5, NP=2, PO=5)}}:
 Index Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   0   |   0   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   0   |   0   0
-----------------------------------------------------------------
   3:   1.0000000000000000e+00      1      0   0   1   |   0   0
julia> Δxc = @vars(d5, complex=true)3-element Vector{ComplexTPS64{Descriptor(NV=3, MO=5, NP=2, PO=5)}}:
 Index Real                     Imag                       Order   Exponent
--------------------------------------------------------------------------------
   1:   1.0000000000000000e+00   0.0000000000000000e+00      1      1   0   0  ⋯
--------------------------------------------------------------------------------
   2:   1.0000000000000000e+00   0.0000000000000000e+00      1      0   1   0  ⋯
--------------------------------------------------------------------------------
   3:   1.0000000000000000e+00   0.0000000000000000e+00      1      0   0   1  ⋯
                                                               3 columns omitted
julia> Δxd = @vars(d5, dynamic=true)3-element Vector{TPS64{GTPSA.Dynamic}}:
Descriptor(NV=3, MO=5, NP=2, PO=5)
 Index Coefficient                Order   Exponent
-----------------------------------------------------------------
   1:   1.0000000000000000e+00      1      1   0   0   |   0   0
-----------------------------------------------------------------
   2:   1.0000000000000000e+00      1      0   1   0   |   0   0
-----------------------------------------------------------------
   3:   1.0000000000000000e+00      1      0   0   1   |   0   0
julia> Δk = @params(d5, complex=true, dynamic=true)2-element Vector{ComplexTPS64{GTPSA.Dynamic}}:
Descriptor(NV=3, MO=5, NP=2, PO=5)
 Index Real                     Imag                       Order   Exponent
--------------------------------------------------------------------------------
   1:   1.0000000000000000e+00   0.0000000000000000e+00      1      0   0   0  ⋯
--------------------------------------------------------------------------------
   2:   1.0000000000000000e+00   0.0000000000000000e+00      1      0   0   0  ⋯
                                                               3 columns omitted

Documentation

GTPSA.@vars — Macro

@vars(descriptor [,complex=bool] [, dynamic=bool])

Constructs a vector of TPSs corresponding to each of the variables in the GTPSA descriptor

Keyword Arguments

complex – If true, returns the variables as ComplexTPS64s. Default is false.

dynamic – If true, the variables will use dynamic Descriptor resolution. Default is false.

GTPSA.@params — Macro

@params(descriptor [,complex=bool] [, dynamic=bool])

Constructs a vector of TPSs corresponding to each of the parameters in the GTPSA descriptor

Keyword Arguments

complex – If true, returns the parameters as ComplexTPS64s. Default is false.

dynamic – If true, the parameters will use dynamic Descriptor resolution. Default is false.