TPS Methods

GTPSA.clear! — Function

clear!(t::Union{TPS,ComplexTPS})

Clears the TPS (sets all monomial coefficients to 0).

GTPSA.complex! — Function

complex!(ct::ComplexTPS, t::TPS)

Sets the pre-allocated ComplexTPS ct equal to t.

GTPSA.compose! — Function

compose!(m::Vector{<:Union{TPS,ComplexTPS}}, m2::Vector{<:Union{TPS,ComplexTPS}}, m1::Vector{<:Union{TPS,ComplexTPS}}; work_low::Union{Nothing,Tuple{Vararg{Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}}}}=nothing, work_prom::Union{Nothing,Tuple{Vararg{Vector{<:ComplexTPS}}}}=nothing)

Composes the vector functions m2 ∘ m1 and stores the result in-place in m. Promotion is allowed, provided the output vector function m has the correct promoted type.

For all compositions, 3 temporary vectors must be generated that contain Ptr{RTPSA} or Ptr{CTPSA} for each TPS in the map (depending on output type), to pass to the low-level C composition function in GTPSA. They are correspondingly referred to as outx_low, m2x_low, and m1x_low. These three temporaries containers can be optionally passed as a tuple in work_low, and must satisfy the following requirements:

work_low[1] = outx_low   # Length >= length(m) = length(m2)
work_low[2] = m2x_low    # Length >= length(m2) = length(m)
work_low[3] = m1x_low    # Length >= length(m1)

If promotion is occuring, then one of the input vectors must be promoted to ComplexTPS. A vector of pre-allocated ComplexTPSs can optionally provided as the first argument in the work_prom tuple, and has the requirement:

If eltype(m.x) != eltype(m1.x) (then m1 must be promoted): work_prom[1] = m1x_prom # Length >= length(m1), Vector{ComplexTPS}

else if eltype(m.x) != eltype(m2.x) (then m2 must be promoted): work_prom[1] = m2x_prom # Length >= length(m2) = length(m), Vector{ComplexTPS}

Note that the ComplexTPSs in the vectors must be allocated and have the same Descriptor.

GTPSA.cutord — Function

cutord(t1::Union{TPS, ComplexTPS}, order::Integer)

Cuts out the monomials in t1 at the given order and above. Or, if order is negative, will cut monomials with orders at and below abs(order).

Examples

julia> d = Descriptor(1,10);

julia> x = vars(d);

julia> cutord(sin(x[1]), 5)
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1
  -1.6666666666666666e-01    3        3

julia> cutord(sin(x[1]), -5)
TPS:
  Coefficient              Order     Exponent
  -1.9841269841269841e-04    7        7
   2.7557319223985893e-06    9        9

GTPSA.cutord! — Function

cutord!(t::T, t1::T, order::Integer) where {T<:Union{TPS,ComplexTPS}}

Cuts out the monomials in t1 at the given order and above. Or, if order is negative, will cut monomials with orders at and below abs(order). t is filled in-place with the result. See the documentation for cutord for examples.

GTPSA.deriv — Function

deriv(t1::Union{TPS,ComplexTPS}, v::Union{TPSIndexType, Nothing}=nothing; param::Union{Integer,Nothing}=nothing, params::Union{SMIndexType, Nothing}=nothing)
∂(t1::Union{TPS,ComplexTPS}, v::Union{TPSIndexType, Nothing}=nothing; param::Union{Integer,Nothing}=nothing, params::Union{SMIndexType, Nothing}=nothing)

Differentiates t1 wrt the variable/parameter specified by the variable/parameter index, or alternatively any monomial specified by indexing-by-order OR indexing-by-sparse monomial.

Input

v – An integer (for variable index), vector/tuple of orders for each variable (for indexing-by-order), or vector/tuple of pairs (sparse monomial)

param – (Keyword argument, optional) An integer for the parameter index

params – (Keyword argument, optional) Vector/tuple of pairs for sparse-monomial indexing

Examples: Variable/Parameter Index:

julia> d = Descriptor(1,5,1,5);

julia> x1 = vars(d)[1]; k1 = params(d)[1];

julia> deriv(x1*k1, 1)
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        0    1

julia> deriv(x1*k1, param=1)
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1    0

Examples: Monomial Index-by-Order

julia> deriv(x1*k1, [1,0])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        0    1

julia> deriv(x1*k1, [0,1])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1    0

julia> deriv(x1*k1, [1,1])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    0        0    0

Examples: Monomial Index-by-Sparse Monomial

julia> deriv(x1*k1, [1=>1])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        0    1

julia> deriv(x1*k1, params=[1=>1])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    1        1    0

julia> deriv(x1*k1, [1=>1], params=[1=>1])
TPS:
  Coefficient              Order     Exponent
   1.0000000000000000e+00    0        0    0

GTPSA.deriv! — Function

deriv!(t::T, t1::T, v::Union{TPSIndexType, Nothing}=nothing; param::Union{Integer,Nothing}=nothing, params::Union{SMIndexType, Nothing}=nothing) where {T<:Union{TPS,ComplexTPS}}
∂!(t::T, t1::T, v::Union{TPSIndexType, Nothing}=nothing; param::Union{Integer,Nothing}=nothing, params::Union{SMIndexType, Nothing}=nothing) where {T<:Union{TPS,ComplexTPS}}

Differentiates t1 wrt the variable/parameter specified by the variable/parameter index, or alternatively any monomial specified by indexing-by-order OR indexing-by-sparse monomial, and sets t equal to the result in-place. See the deriv documentation for examples.

Input

v – An integer (for variable index), vector/tuple of orders for each variable (for indexing-by-order), or vector/tuple of pairs (sparse monomial)

param – (Keyword argument, optional) An integer for the parameter index

params – (Keyword argument, optional) Vector/tuple of pairs for sparse-monomial indexing

GTPSA.evaluate — Function

evaluate(m::Vector{T}, tb::Vector{<:Number}) where {T<:Union{TPS,ComplexTPS}}

Evaluates the vector function m at the point tb.

GTPSA.evaluate! — Function

evaluate!(tc::Vector{<:Number}, m::Vector{T}, tb::Vector{<:Number}; work_low::Vector{<:Union{Ptr{RTPSA},Ptr{CTPSA}}}=Vector{lowtype(T)}(undef, length(m))) where {T<:Union{TPS,ComplexTPS}}

Evaluates the vector function m at the point tb, and fills tc with the result. An optional container work_low can be provided for containing the low-level TPS structs for zero allocations.

GTPSA.getord — Function

getord(t1::Union{TPS, ComplexTPS}, order::Integer)

Extracts one homogenous polynomial from t1 of the given order.

GTPSA.getord! — Function

getord!(t::T, t1::T, order::Integer) where {T<:Union{TPS,ComplexTPS}}

Extracts one homogenous polynomial from t1 of the given order and fills t with the result in-place.

GTPSA.integ — Function

integ(t1::Union{TPS, ComplexTPS}, var::Integer=1)
∫(t1::Union{TPS, ComplexTPS}, var::Integer=1)

Integrates t1 wrt the variable var. Integration wrt parameters is not allowed, and integration wrt higher order monomials is not currently supported.

GTPSA.integ! — Function

integ!(t::T, t1::T, var::Integer=1) where {T<:Union{TPS,ComplexTPS}}
∫!(t::T, t1::T, var::Integer=1) where {T<:Union{TPS,ComplexTPS}}

Integrates t1 wrt the variable var and fills t with the result. Integration wrt parameters is not allowed, and integration wrt higher order monomials is not currently supported.

GTPSA.scalar — Function

scalar(t::Union{TPS,ComplexTPS})

Extracts the scalar part of the TPS. Equivalent to t[0] but this can be easily broadcasted.

GTPSA.translate — Function

Returns a vector function equal to ma with its expansion point translated by tb

GTPSA.translate! — Function

translate!(mc::Vector{<:T}, ma::Vector{<:T}, tb::Vector{<:Number}) where {T<:Union{TPS,ComplexTPS}}

Fills ma with the vector function equal to ma with its expansion point translated by tb.

Two temporary vectors of either Ptr{RTPSA} or Ptr{CTPSA} must be created, or they can optionally be passed as a tuple to the kwarg work_low where

ma_low = low corresponding to ma = work_low[1] mb_low = low corresponding to mb = work_low[2]